Add the create book and label method
